Initial refactoring of lokqldx to provide easier tracking against lokqldx-av
