+++ applications/lokqlDx-Av/ResourceHelper.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Reflection;
 
 namespace LokqlDx;
+
 internal static class ResourceHelper
 {
     /// <summary>
