+++ applications/lokqlDx/Views/Dialogs/ApplicationPreferencesWindow.xaml
