+++ applications/lokqlDx/Views/Dialogs/ApplicationPreferencesWindow.xaml.cs
