+++ applications/lokqlDx/Views/Dialogs/EnterTitle.xaml
