+++ applications/lokqlDx/Views/Dialogs/MarkdownHelpWindow.xaml.cs
