+++ applications/lokqlDx/Views/Dialogs/WorkspacePreferencesWindow.xaml.cs
