+++ applications/lokqlDx/Views/QueryEditor.xaml
