+++ applications/lokqlDx/Views/QueryEditor.xaml.cs
@@ -17,6 +17,7 @@ using ICSharpCode.AvalonEdit.Search;
 using Intellisense;
 using KustoLoco.Core.Settings;
 using Lokql.Engine;
+using LokqlDx;
 using NotNullStrings;
 using FontFamily = System.Windows.Media.FontFamily;
 namespace lokqlDx;
@@ -233,31 +234,14 @@ public partial class QueryEditor : UserControl
         }
     }
 
-    /// <summary>
-    ///     Gets a resource name independent of namespace
-    /// </summary>
-    /// <remarks>
-    ///     For some reason dotnet publish decides to lower-case the
-    ///     namespace in the resource name. In any case, we really don't want to trust
-    ///     that the namespace won't change so do a match against the filename
-    /// </remarks>
-    private Stream SafeGetResourceStream(string substring)
-    {
-        var assembly = Assembly.GetExecutingAssembly();
-        var availableResources = assembly.GetManifestResourceNames();
-        var wanted =
-            availableResources.Single(name => name.Contains(substring, StringComparison.CurrentCultureIgnoreCase));
-        return assembly.GetManifestResourceStream(wanted)!;
-    }
-
     private void QueryEditor_OnLoaded(object sender, RoutedEventArgs e)
     {
-        using var s = SafeGetResourceStream("SyntaxHighlighting.xml");
+        using var s = ResourceHelper.SafeGetResourceStream("SyntaxHighlighting.xml");
         using var reader = new XmlTextReader(s);
         Query.SyntaxHighlighting = HighlightingLoader.Load(reader, HighlightingManager.Instance);
-        using var functions = SafeGetResourceStream("IntellisenseFunctions.json");
+        using var functions = ResourceHelper.SafeGetResourceStream("IntellisenseFunctions.json");
         _kqlFunctionEntries = JsonSerializer.Deserialize<IntellisenseEntry[]>(functions!)!;
-        using var ops = SafeGetResourceStream("IntellisenseOperators.json");
+        using var ops = ResourceHelper.SafeGetResourceStream("IntellisenseOperators.json");
         KqlOperatorEntries = JsonSerializer.Deserialize<IntellisenseEntry[]>(ops!)!;
         SearchPanel.Install(Query);
     }
