+++ applications/lokqlDx/Workspaces/WorkspaceManager.cs
